[fix 🎭] fix code & setup uploadthing file uploader
