+++ .env.exmaple
@@ -1,3 +1,5 @@
 DATABASE_URI=""
 
-
+# Upload Things
+UPLOADTHING_SECRET=""
+UPLOADTHING_APP_ID=""
