--- src/components/providers/Providers.tsx
@@ -1,24 +1,0 @@
-"use client";
-
-import React, { ReactNode } from "react";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { httpBatchLink } from "@trpc/client";
-import { getBaseUrl, trpc } from "@/tRPC/client";
-
-const Providers = ({ children }: { children: ReactNode }) => {
-  const queryClient = new QueryClient();
-  const trpcClient = trpc.createClient({
-    links: [
-      httpBatchLink({
-        url: `${getBaseUrl()}/api/trpc`,
-      }),
-    ],
-  });
-  return (
-    <trpc.Provider queryClient={queryClient} client={trpcClient}>
-      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-    </trpc.Provider>
-  );
-};
-
-export default Providers;
