+++ src/utils/uploadthing.ts
@@ -1,0 +1,9 @@
+// Resource: https://docs.uploadthing.com/api-reference/react#generatereacthelpers
+// Copy paste (be careful with imports)
+
+import { generateReactHelpers } from "@uploadthing/react/hooks";
+
+import type { OurFileRouter } from "@/app/api/uploadthing/core";
+
+export const { useUploadThing, uploadFiles } =
+  generateReactHelpers<OurFileRouter>();
