--- src/validators/index.ts
@@ -1,1 +1,0 @@
-import {} from "zod";
